fixed employee dto and added build.gradle

+++ src/main/java/com/LogisticsCompany/dto/EmployeeDTO.java
@@ -5,14 +5,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigDecimal;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class EmployeeDTO {
     private String name;
-    private BigDecimal salary;
+    private double salary;
     private Long officeID;
     private EmployeeType employeeType;
 }
